move and fix test
